feat!: multiple modifications to the original SILA method
BREAKING CHANGE: the function sila is no longer needed and is removed.
BREAKING CHANGE: illa and sila_estimate function calls have been simplified.

Smoothing of the sampled rates vs. value is now donw using a generalized
additive model with a log-link to ensure non-negativity

Integration is performed using the lsoda ODE integrator

Default approach for estimating time shifts is to use all visits per subject
rather than the last visit only

An initial check examines if biomarkers are globally increasing or decreasing
with age (decreasing is not currently supported)

A warning is displayed if the monotonicity assumption may be violated